+++ extensions/mongodb/src/test/java/com/hazelcast/jet/mongodb/dataconnection/MongoDataConnectionSslTest.java
@@ -15,76 +15,28 @@
  */
 package com.hazelcast.jet.mongodb.dataconnection;
 
-import com.github.dockerjava.api.command.InspectContainerResponse;
 import com.hazelcast.config.DataConnectionConfig;
-import com.hazelcast.jet.SimpleTestInClusterSupport;
 import com.hazelcast.jet.test.IgnoreInJenkinsOnWindows;
 import com.hazelcast.test.HazelcastSerialClassRunner;
 import com.hazelcast.test.annotation.ParallelJVMTest;
 import com.hazelcast.test.annotation.QuickTest;
-import com.mongodb.ConnectionString;
-import com.mongodb.MongoClientSettings;
-import com.mongodb.client.MongoClient;
-import com.mongodb.client.MongoClients;
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
-import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.testcontainers.containers.MongoDBContainer;
-import org.testcontainers.containers.output.Slf4jLogConsumer;
 
-import java.io.File;
-import java.io.IOException;
-import java.io.InputStream;
-import java.nio.file.Files;
-import java.nio.file.StandardCopyOption;
-import java.util.ArrayList;
-
-import static com.hazelcast.jet.impl.util.ExceptionUtil.rethrow;
-import static com.hazelcast.jet.mongodb.AbstractMongoTest.TEST_MONGO_VERSION;
-import static com.hazelcast.jet.mongodb.impl.Mappers.defaultCodecRegistry;
-import static com.hazelcast.test.DockerTestUtil.assumeDockerEnabled;
-import static org.testcontainers.containers.BindMode.READ_WRITE;
+import javax.annotation.Nonnull;
+import javax.net.ssl.SSLContext;
 
 @RunWith(HazelcastSerialClassRunner.class)
 @Category({QuickTest.class, ParallelJVMTest.class, IgnoreInJenkinsOnWindows.class})
-public class MongoDataConnectionSslTest extends SimpleTestInClusterSupport {
+public class MongoDataConnectionSslTest extends MongoDataConnectionSslTestBase {
 
-    private static final String DATABASE = "MongoDataConnectionSslTest";
-    private static final String TEST_COLLECTION = "testCollection";
-    private static final Logger LOGGER = LoggerFactory.getLogger(MongoDataConnectionSslTest.class);
-    private static final MongoDBContainer mongoContainer = new MyMongoContainer()
-            .withEnv("MONGO_INITDB_DATABASE", DATABASE)
-            .withClasspathResourceMapping("certs/localhost.pem", "/data/ssl/key.pem", READ_WRITE)
-            .withClasspathResourceMapping("tlsMongo.yaml", "/etc/mongo/mongod.conf", READ_WRITE)
-            .withExposedPorts(27017)
-            .withLogConsumer(new Slf4jLogConsumer(LOGGER))
-            .withCommand("--config", "/etc/mongo/mongod.conf");
-    private static String connectionString;
-    private static String trustStoreLocation;
-
-    @BeforeClass
-    public static void setUp() {
-        try (InputStream resourceTS = MongoDataConnectionSslTest.class.getResourceAsStream("/certs/ca.p12")) {
-            File tempFileTS = File.createTempFile("MongoDataConnectionSslTest", "jks");
-            Files.copy(resourceTS, tempFileTS.toPath(), StandardCopyOption.REPLACE_EXISTING);
-
-            System.setProperty("https.protocols", "TLSv1,TLSv1.1,TLSv1.2,TLSv1.3");
-            System.setProperty("javax.net.debug", "ssl:handshake");
-            trustStoreLocation = tempFileTS.toString();
-        } catch (IOException e) {
-            throw rethrow(e);
-        }
-        assumeDockerEnabled();
-        mongoContainer.start();
-        connectionString = "mongodb://" + mongoContainer.getHost() + ":" + mongoContainer.getMappedPort(27017)
-        + "/";
+    @Override
+    protected String getConfigurationFile() {
+        return "tlsMongo.yaml";
     }
 
-    private static DataConnectionConfig getDataConnectionConfig() {
+    @Override
+    protected DataConnectionConfig getDataConnectionConfig() {
         return new DataConnectionConfig("mongoDB")
                 .setType("Mongo")
                 .setName("mongoDB")
@@ -98,52 +50,11 @@ public class MongoDataConnectionSslTest extends SimpleTestInClusterSupport {
                 .setShared(true);
     }
 
-    @AfterClass
-    public static void clear() {
-        if (mongoContainer != null) {
-            mongoContainer.stop();
-        }
-    }
-
-    @Test
-    public void should_connect_with_ssl() {
-        // given / when
-        try (MongoClient mongo = mongoClient()) {
-            mongo.getDatabase(DATABASE).createCollection(TEST_COLLECTION);
-        }
-        MongoDataConnection dataConnection = new MongoDataConnection(getDataConnectionConfig());
-        MongoClient client = dataConnection.getClient();
-
-        // then
-        client.getDatabase(DATABASE).getCollection(TEST_COLLECTION).find().into(new ArrayList<>());
-        // no exception should be thrown
-        dataConnection.destroy();
-    }
-
-    private MongoClient mongoClient() {
-        return MongoClients.create(MongoClientSettings
-                .builder()
-                .applyConnectionString(new ConnectionString(connectionString))
-                .applyToSslSettings(b -> {
-                    b.enabled(true).invalidHostNameAllowed(true);
-                    b.context(SslConf.createSSLContext(null, null, null,
-                            trustStoreLocation, "pkcs12", "123456".toCharArray()));
-                })
-                .codecRegistry(defaultCodecRegistry())
-                .build());
-    }
-
-    private static class MyMongoContainer extends MongoDBContainer {
-        MyMongoContainer() {
-            super("mongo:" + TEST_MONGO_VERSION);
-        }
-        @Override
-        public void configure() {
-        }
-
-        @Override
-        protected void containerIsStarted(InspectContainerResponse containerInfo, boolean reused) {
-        }
+    @Nonnull
+    @Override
+    protected SSLContext getSslContext() {
+        return SslConf.createSSLContext(null, null, null,
+                trustStoreLocation, "pkcs12", "123456".toCharArray());
     }
 
 }
